+++ src/colab/client.ts
@@ -8,7 +8,14 @@ import { UUID } from 'crypto';
 import * as https from 'https';
 import fetch, { Headers, Request, RequestInit, Response } from 'node-fetch';
 import { z } from 'zod';
+import { fetchAndParse } from '../common/fetch-utils';
 import { traceMethod } from '../common/logging/decorators';
+import {
+  buildFetchChain,
+  createAcceptJsonMiddleware,
+  createAuthMiddleware,
+  createErrorMiddleware,
+} from '../common/middleware';
 import { Session } from '../jupyter/client/generated';
 import { ColabAssignedServer } from '../jupyter/servers';
 import { uuidToWebSafeBase64 } from '../utils/uuid';
@@ -41,7 +48,6 @@ import {
   ResourcesSchema,
 } from './api';
 import { ColabRequestError } from './errors';
-import { fetchAndParse } from './fetch-utils';
 import {
   COLAB_CLIENT_AGENT_HEADER,
   COLAB_RUNTIME_PROXY_TOKEN_HEADER,
@@ -50,12 +56,6 @@ import {
   COLAB_VS_CODE_EXTENSION_VERSION,
   COLAB_XSRF_TOKEN_HEADER,
 } from './headers';
-import {
-  buildFetchChain,
-  createAcceptJsonMiddleware,
-  createAuthMiddleware,
-  createErrorMiddleware,
-} from './middleware';
 
 const TUN_ENDPOINT = '/tun/m';
