+++ src/common/middleware.ts
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import { Request, RequestInit, Response } from 'node-fetch';
-import { ColabRequestError } from './errors';
-import { ACCEPT_JSON_HEADER, AUTHORIZATION_HEADER } from './headers';
+import { ColabRequestError } from '../colab/errors';
+import { ACCEPT_JSON_HEADER, AUTHORIZATION_HEADER } from '../colab/headers';
 
 /**
  * A middleware function for fetch requests. It takes a Request and a next
