+++ src/common/middleware.unit.test.ts
@@ -6,7 +6,7 @@
 import { expect } from 'chai';
 import { Request, Response } from 'node-fetch';
 import * as sinon from 'sinon';
-import { ColabRequestError } from './errors';
+import { ColabRequestError } from '../colab/errors';
 import {
   FetchMiddleware,
   buildFetchChain,
